pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "swarm"
version = "0.1.0"
description = "A lightweight, stateless multi-agent orchestration framework."
authors = ["OpenAI Solutions"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12.5"
numpy = "*"
openai = ">=1.33.0"
pytest = "*"
requests = "*"
tqdm = "*"
pre-commit = "*"
instructor = "*"
pillow = "^11.0.0"

[tool.autopep8]
max_line_length = 120
ignore = "E501,W6"
in-place = true
recursive = true
aggressive = 3